% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkin_out.R
\name{spcheckout}
\alias{spcheckout}
\alias{spundocheckout}
\alias{spcheckin}
\title{Check-in and Check-out Files}
\usage{
spcheckout(file, directory, site, site_collection)

spundocheckout(file, directory, site, site_collection)

spcheckin(comment, file, directory, site, site_collection)
}
\arguments{
\item{file}{a name of a file on sharepoint.}

\item{directory}{a directory name.}

\item{site}{a SharePoint site name, e.g. '/ExpertGroups/WGNSSK'.}

\item{site_collection}{a SharePoint site collection, will almost exclusively
be 'https://community.ices.dk', so these functions should work for other
SharePoint sites outside ICES.}

\item{comment}{the checkin comment}
}
\value{
invisibly returns the response from the SharePoint server
}
\description{
Check out the file first to prevent other users from changing it.
Then, check it back in after you've made your changes.
}
\seealso{
\code{\link{spfile.create}} To upload a file.
}
