% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{spdir.create}
\alias{spdir.create}
\alias{spdir.delete}
\title{Create and delete folders}
\usage{
spdir.create(folder, directory, site, site_collection)

spdir.delete(folder, directory, site, site_collection)
}
\arguments{
\item{folder}{a name of a folder on sharepoint.}

\item{directory}{a directory name.}

\item{site}{a SharePoint site name, e.g. '/ExpertGroups/WGNSSK'.}

\item{site_collection}{a SharePoint site collection, will almost exclusively}
}
\value{
invisibly returns the response from the SharePoint server
}
\description{
Create and delete folders on the SharePoint system.
}
\seealso{
\code{\link{spfile.create}, \link{spfile.delete}} To upload a file and delete files.
}
