% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file.R
\name{spfile.create}
\alias{spfile.create}
\alias{spfile.delete}
\title{Create and delete files}
\usage{
spfile.create(file, con, directory, site, site_collection, text = "")

spfile.delete(file, directory, site, site_collection)
}
\arguments{
\item{file}{a name of a file on sharepoint.}

\item{con}{a file connection to, or the file name of, the file to upload.}

\item{directory}{a directory name.}

\item{site}{a SharePoint site name, e.g. '/ExpertGroups/WGNSSK'.}

\item{site_collection}{a SharePoint site collection, will almost exclusively
be 'https://community.ices.dk', so these functions should work for other
SharePoint sites outside ICES.}

\item{text}{if `con` is not supplied, the `text` is assummed to be the contents
of the file, defaulting to "".}
}
\value{
invisibly returns the response from the SharePoint server
}
\description{
Create and delete files on the SharePoint system.
}
\seealso{
\code{\link{spcheckout}} To upload a file.
}
