% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSharePoint-package.R
\docType{package}
\name{icesSharePoint-package}
\alias{icesSharePoint-package}
\alias{icesSharePoint}
\title{Helper functions to interact with the ICES sharepoint.}
\description{
Helper functions to download files from the ICES sharepoint.
}
\references{
ICES Sharepoint site: \url{https://community.ices.dk/SitePages/Home.aspx}.
}
\author{
Colin Millar and Arni Magnusson.
}
