% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setget_wd_and_site.R
\name{setspwd}
\alias{setspwd}
\alias{getspwd}
\alias{setspsite}
\alias{getspsite}
\title{Working Directory and Site}
\usage{
setspwd(directory)

getspwd()

setspsite(site)

getspsite()
}
\arguments{
\item{directory}{a directory name.}

\item{site}{a SharePoint site name, e.g. '/ExpertGroups/WGNSSK'.}
}
\value{
getspwd returns an absolute filepath representing the current SharePoint working directory
within the current site; setspwd(directory) is used to set the working directory to directory.
}
\description{
Set and get SharePoint working directory and site.
}
