% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{spgetfile}
\alias{spgetfile}
\title{Download Files from SharePoint}
\usage{
spgetfile(file, site, site_collection, destdir = dirname(file))
}
\arguments{
\item{file}{a name of a file on sharepoint.}

\item{site}{a SharePoint site name, e.g. '/ExpertGroups/WGNSSK'.}

\item{site_collection}{a SharePoint site collection, will almost exclusively
be 'https://community.ices.dk', so these functions should work for other
SharePoint sites outside ICES.}

\item{destdir}{a relative path of the directory to download the file to.}
}
\value{
The path of the downloaded file
}
\description{
Download file from sharepoint and return file location.
}
