% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info.R
\name{spgetuser}
\alias{spgetuser}
\title{Get Current User Information}
\usage{
spgetuser(site, site_collection)
}
\arguments{
\item{site}{a SharePoint site name, e.g. '/ExpertGroups/WGNSSK'.}

\item{site_collection}{a SharePoint site collection, will almost exclusively
be 'https://community.ices.dk', so these functions should work for other
SharePoint sites outside ICES.}

\item{dir}{a relative path of the directory to download the file to.}
}
\value{
The path of the downloaded file
}
\description{
Download file from sharepoint and return file location.
}
