% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{spservice}
\alias{spservice}
\title{REST API request from SharePoint}
\usage{
spservice(service, site, site_collection)
}
\arguments{
\item{service}{an api request.}

\item{site}{a SharePoint site name, e.g. '/ExpertGroups/WGNSSK'. The default
is to take the value of getOption("icesSharePoint.site").}

\item{site_collection}{a SharePoint site collection, will almost exclusively
be 'https://community.ices.dk', so these functions should work for other
SharePoint sites outside ICES. The default
is to take the value of getOption("icesSharePoint.site_collection").}
}
\value{
The result of the request.
}
\description{
Implement a REST API request.  This function wraps up an API request into
single uri and calls it
}
\details{
The requests always ask for json data, but may on occasion (esp when
the request fails) return xml.
}
\examples{
\dontrun{
 x <- spservice("sitegroups/getbyname('ICES\%20staff')/Users")
cbind(sapply(x$results, "[[", "Title"),
      sapply(x$results, "[[", "Email"))
}
}
