% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{sppost}
\alias{sppost}
\title{Basic POST request from SharePoint}
\usage{
sppost(uri, ...)
}
\arguments{
\item{uri}{an api request.}
}
\value{
The result of the request.
}
\description{
Implement a POST request for a given uri.
}
\details{
The requests always ask for json data, but may on occasion (esp when
the request fails) return xml.
}
